+++ src/hotelling.jl
@@ -1,14 +1,14 @@
 """
-    hotelling(X, y::Vector{String}; digits::Int = 5)
+    hotelling(X::AbstMatVec{Q}, y::Vector{String}; digits::Int = 5) where Q <: Float
 Two-sample Hotelling's T-squared test.
-* `X` : X-data (n, p).
+* `X` : A matrix (n, p) or vector (n).
 * `y` : A categorical variable (2-class membership) (n). Must be a `Vector{String}`.
 Keyword arguments:
 * `digits` : Nb. digits for the outputs.
 
 ## Examples 
 ```julia
-using Jchemo, JchemoData, JLD2
+using Jchemo, JchemoData, JLD2, StatsModels
 path_jdat = dirname(dirname(pathof(JchemoData)))
 db = joinpath(path_jdat, "data/reaction_bertinetto.jld2")
 @load db dat
@@ -17,7 +17,7 @@ datf = dat.datf
 n = nro(datf)
 tab(datf; group = [:temp, :catal])  # balanced design
 ##
-Y = datf[:, [:y1, :y2]]
+Y = Matrix(datf[:, [:y1, :y2]])
 fact = datf.catal
 tab(fact)
 
@@ -33,7 +33,7 @@ W = matW(Y, fact, pweight(ones(n))).W
 manova(Y, @formula(0 ~ catal), datf; test)
 ```
 """
-function hotelling(X, y::Vector{String}; digits::Int = 5)
+function hotelling(X::AbstMatVec{Q}, y::Vector{String}; digits::Int = 5) where Q <: Float
     n, p = size(X)
     res = matWc(X, y)
     ni = res.ni