+++ src/matW.jl
@@ -1,7 +1,7 @@
 """
-    matB(X::Matrix{Q}, y::Vector{String}, weights::ProbabilityWeights{Q}) where Q <: Float
+    matB(X::AbstMatVec{Q}, y::Vector{String}, weights::ProbabilityWeights{Q}) where Q <: Float
 Between-class covariance matrix.
-* `X` : X-matrix (n, p).
+* `X` : A matrix (n, p) or vector (n).
 * `y` : A categorical variable (class membership) (n). Must be a `Vector{String}`.
 * `weights` : Weights (n) of the observations. Must be of type `ProbabilityWeights` (see e.g., function `pweight`).
 
@@ -41,7 +41,7 @@ matW(X, y, weights).W + matB(X, y, weights).B
 covm(X, weights)
 ```
 """ 
-function matB(X::Matrix{Q}, y::Vector{String}, weights::ProbabilityWeights{Q}) where Q <: Float
+function matB(X::AbstMatVec{Q}, y::Vector{String}, weights::ProbabilityWeights{Q}) where Q <: Float
     p = nco(X)
     taby = tab(y)
     lev = taby.keys
@@ -58,9 +58,9 @@ function matB(X::Matrix{Q}, y::Vector{String}, weights::ProbabilityWeights{Q}) w
 end
 
 """
-    matW(X::Matrix{Q}, y::Vector{String}, weights::ProbabilityWeights{Q}) where Q <: Float
+    matW(X::AbstMatVec{Q}, y::Vector{String}, weights::ProbabilityWeights{Q}) where Q <: Float
 Within-class (non-corrected) covariance matrices.
-* `X` : X-matrix (n, p).
+* `X` : A matrix (n, p) or vector (n).
 * `y` : A categorical variable (class membership) (n). Must be a `Vector{String}`.
 * `weights` : Weights (n) of the observations. Must be of type `ProbabilityWeights` (see e.g., function `pweight`).
 
@@ -71,7 +71,7 @@ If class i contains only one observation, `Wi` is computed by:
 
 For examples, see function `matB`. 
 """ 
-function matW(X::Matrix{Q}, y::Vector{String}, weights::ProbabilityWeights{Q}) where Q <: Float
+function matW(X::AbstMatVec{Q}, y::Vector{String}, weights::ProbabilityWeights{Q}) where Q <: Float
     p = nco(X) 
     taby = tab(y)
     lev = taby.keys
@@ -102,9 +102,9 @@ function matW(X::Matrix{Q}, y::Vector{String}, weights::ProbabilityWeights{Q}) w
 end
 
 """
-    matWc(X::Matrix{Q}, y::Vector{String}) where Q <: Float
+    matWc(X::AbstMatVec{Q}, y::Vector{String}) where Q <: Float
 Within-class (corrected) covariance matrices.
-* `X` : X-matrix (n, p).
+* `X` : A matrix (n, p) or vector (n).
 * `y` : A categorical variable (class membership) (n). Must be a `Vector{String}`.
 
 Compute the (corrected) within-class and pooled covariance matrices (outputs `Wi` and `W`, respectively) of `X`.